refactor: extract filter / evaluator
